requirements.txt
fastapi
uvicorn[standard]

qdrant-client

langchain
langchain-community
langchain-openai

python-dotenv
pydantic
